[fold] Mv STXChainAttestationBatch functions to XChainAttestations file https://github.com/XRPLF/rippled/pull/4292#pullrequestreview-1308324738
