+++ src/ripple/protocol/XChainAttestations.h
@@ -22,23 +22,250 @@
 
 #include <ripple/basics/Buffer.h>
 #include <ripple/basics/Expected.h>
+#include <ripple/ledger/ReadView.h>
 #include <ripple/protocol/AccountID.h>
 #include <ripple/protocol/Issue.h>
 #include <ripple/protocol/PublicKey.h>
 #include <ripple/protocol/SField.h>
-#include <ripple/protocol/STXChainAttestationBatch.h>
+#include <ripple/protocol/STBase.h>
 #include <ripple/protocol/STXChainBridge.h>
+#include <ripple/protocol/SecretKey.h>
 #include <ripple/protocol/TER.h>
 
+#include <boost/container/flat_set.hpp>
+#include <boost/container/vector.hpp>
+
 #include <cstddef>
+#include <type_traits>
+#include <unordered_map>
 #include <vector>
 
 namespace ripple {
 
-namespace AttestationBatch {
-struct AttestationClaim;
-struct AttestationCreateAccount;
-}  // namespace AttestationBatch
+namespace Attestations {
+
+// Check that the public key is allowed to sign for the given account. If the
+// account does not exist on the ledger, then the public key must be the master
+// key for the given account if it existed. Otherwise the key must be an enabled
+// master key or a regular key for the existing account.
+TER
+checkAttestationPublicKey(
+    ReadView const& view,
+    std::unordered_map<AccountID, std::uint32_t> const& signersList,
+    AccountID const& attestationSignerAccount,
+    PublicKey const& pk,
+    beast::Journal j);
+
+struct AttestationBase
+{
+    // Account associated with the public key
+    AccountID attestationSignerAccount;
+    // Public key from the witness server attesting to the event
+    PublicKey publicKey;
+    // Signature from the witness server attesting to the event
+    Buffer signature;
+    // Account on the sending chain that triggered the event (sent the
+    // transaction)
+    AccountID sendingAccount;
+    // Amount transfered on the sending chain
+    STAmount sendingAmount;
+    // Account on the destination chain that collects a share of the attestation
+    // reward
+    AccountID rewardAccount;
+    // Amount was transfered on the locking chain
+    bool wasLockingChainSend;
+
+    explicit AttestationBase(
+        AccountID attestationSignerAccount_,
+        PublicKey const& publicKey_,
+        Buffer signature_,
+        AccountID const& sendingAccount_,
+        STAmount const& sendingAmount_,
+        AccountID const& rewardAccount_,
+        bool wasLockingChainSend_);
+
+    AttestationBase(AttestationBase const&) = default;
+
+    virtual ~AttestationBase() = default;
+
+    AttestationBase&
+    operator=(AttestationBase const&) = default;
+
+    // verify that the signature attests to the data.
+    bool
+    verify(STXChainBridge const& bridge) const;
+
+protected:
+    explicit AttestationBase(STObject const& o);
+    explicit AttestationBase(Json::Value const& v);
+
+    [[nodiscard]] static bool
+    equalHelper(AttestationBase const& lhs, AttestationBase const& rhs);
+
+    [[nodiscard]] static bool
+    sameEventHelper(AttestationBase const& lhs, AttestationBase const& rhs);
+
+    void
+    addHelper(STObject& o) const;
+
+private:
+    [[nodiscard]] virtual std::vector<std::uint8_t>
+    message(STXChainBridge const& bridge) const = 0;
+};
+
+// Attest to a regular cross-chain transfer
+struct AttestationClaim : AttestationBase
+{
+    std::uint64_t claimID;
+    std::optional<AccountID> dst;
+
+    explicit AttestationClaim(
+        AccountID attestationSignerAccount_,
+        PublicKey const& publicKey_,
+        Buffer signature_,
+        AccountID const& sendingAccount_,
+        STAmount const& sendingAmount_,
+        AccountID const& rewardAccount_,
+        bool wasLockingChainSend_,
+        std::uint64_t claimID_,
+        std::optional<AccountID> const& dst_);
+
+    explicit AttestationClaim(
+        STXChainBridge const& bridge,
+        AccountID attestationSignerAccount_,
+        PublicKey const& publicKey_,
+        SecretKey const& secretKey_,
+        AccountID const& sendingAccount_,
+        STAmount const& sendingAmount_,
+        AccountID const& rewardAccount_,
+        bool wasLockingChainSend_,
+        std::uint64_t claimID_,
+        std::optional<AccountID> const& dst_);
+
+    explicit AttestationClaim(STObject const& o);
+    explicit AttestationClaim(Json::Value const& v);
+
+    [[nodiscard]] STObject
+    toSTObject() const;
+
+    // return true if the two attestations attest to the same thing
+    [[nodiscard]] bool
+    sameEvent(AttestationClaim const& rhs) const;
+
+    [[nodiscard]] static std::vector<std::uint8_t>
+    message(
+        STXChainBridge const& bridge,
+        AccountID const& sendingAccount,
+        STAmount const& sendingAmount,
+        AccountID const& rewardAccount,
+        bool wasLockingChainSend,
+        std::uint64_t claimID,
+        std::optional<AccountID> const& dst);
+
+    [[nodiscard]] bool
+    validAmounts() const;
+
+private:
+    [[nodiscard]] std::vector<std::uint8_t>
+    message(STXChainBridge const& bridge) const override;
+
+    friend bool
+    operator==(AttestationClaim const& lhs, AttestationClaim const& rhs);
+};
+
+struct CmpByClaimID
+{
+    bool
+    operator()(AttestationClaim const& lhs, AttestationClaim const& rhs) const
+    {
+        return lhs.claimID < rhs.claimID;
+    }
+};
+
+// Attest to a cross-chain transfer that creates an account
+struct AttestationCreateAccount : AttestationBase
+{
+    // createCount on the sending chain. This is the value of the `CreateCount`
+    // field of the bridge on the sending chain when the transaction was
+    // executed.
+    std::uint64_t createCount;
+    // Account to create on the destination chain
+    AccountID toCreate;
+    // Total amount of the reward pool
+    STAmount rewardAmount;
+
+    explicit AttestationCreateAccount(STObject const& o);
+
+    explicit AttestationCreateAccount(Json::Value const& v);
+
+    explicit AttestationCreateAccount(
+        AccountID attestationSignerAccount_,
+        PublicKey const& publicKey_,
+        Buffer signature_,
+        AccountID const& sendingAccount_,
+        STAmount const& sendingAmount_,
+        STAmount const& rewardAmount_,
+        AccountID const& rewardAccount_,
+        bool wasLockingChainSend_,
+        std::uint64_t createCount_,
+        AccountID const& toCreate_);
+
+    explicit AttestationCreateAccount(
+        STXChainBridge const& bridge,
+        AccountID attestationSignerAccount_,
+        PublicKey const& publicKey_,
+        SecretKey const& secretKey_,
+        AccountID const& sendingAccount_,
+        STAmount const& sendingAmount_,
+        STAmount const& rewardAmount_,
+        AccountID const& rewardAccount_,
+        bool wasLockingChainSend_,
+        std::uint64_t createCount_,
+        AccountID const& toCreate_);
+
+    [[nodiscard]] STObject
+    toSTObject() const;
+
+    // return true if the two attestations attest to the same thing
+    [[nodiscard]] bool
+    sameEvent(AttestationCreateAccount const& rhs) const;
+
+    friend bool
+    operator==(
+        AttestationCreateAccount const& lhs,
+        AttestationCreateAccount const& rhs);
+
+    [[nodiscard]] static std::vector<std::uint8_t>
+    message(
+        STXChainBridge const& bridge,
+        AccountID const& sendingAccount,
+        STAmount const& sendingAmount,
+        STAmount const& rewardAmount,
+        AccountID const& rewardAccount,
+        bool wasLockingChainSend,
+        std::uint64_t createCount,
+        AccountID const& dst);
+
+    [[nodiscard]] bool
+    validAmounts() const;
+
+private:
+    [[nodiscard]] std::vector<std::uint8_t>
+    message(STXChainBridge const& bridge) const override;
+};
+
+struct CmpByCreateCount
+{
+    bool
+    operator()(
+        AttestationCreateAccount const& lhs,
+        AttestationCreateAccount const& rhs) const
+    {
+        return lhs.createCount < rhs.createCount;
+    }
+};
+
+};  // namespace Attestations
 
 // Result when checking when two attestation match.
 enum class AttestationMatch {
