+++ src/test/jtx/impl/attester.cpp
@@ -21,9 +21,9 @@
 
 #include <ripple/basics/Slice.h>
 #include <ripple/protocol/PublicKey.h>
-#include <ripple/protocol/STXChainAttestationBatch.h>
 #include <ripple/protocol/STXChainBridge.h>
 #include <ripple/protocol/SecretKey.h>
+#include <ripple/protocol/XChainAttestations.h>
 
 namespace ripple {
 namespace test {
